backend/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "redline-ai"
version = "0.1.0"
description = "Emergency Response Intelligence Platform"
authors = [
    {name = "Redline AI Team"}
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.1.0",
    "redis>=5.0.1",
    "sqlalchemy[asyncio]>=2.0.23",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.6",
    "python-json-logger>=2.0.7",
    "structlog>=24.1.0",
    "prometheus-client>=0.20.0",
    "starlette-prometheus>=0.10.0",
    "slowapi>=0.1.9",
    "twilio>=9.0.0",
    "jinja2>=3.1.0",
    "aiofiles>=23.0.0",
    "pybreaker>=1.0.0",
    "filelock>=3.13.0",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
    "respx>=0.20.0",
    "aiosqlite>=0.20.0",
    "groq>=0.4.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "S", "B", "C4", "SIM", "RUF"]
ignore = [
    "B008",    # function-call-in-default-argument (FastAPI Depends pattern)
    "B904",    # raise-without-from (FastAPI HTTPException pattern)
    "RUF001",  # ambiguous unicode in strings
    "RUF002",  # ambiguous unicode in docstrings
    "RUF003",  # ambiguous unicode in comments
    "RUF012",  # mutable class default (SQLAlchemy models)
    "RUF013",  # implicit-optional (legacy pattern)
    "UP042",   # replace-str-enum
    "E501",    # line too long (handled by formatter)
    "E721",    # type-comparison (schema checks in tests)
    "C408",    # unnecessary-collection-call
    "RUF006",  # asyncio-dangling-task (fire-and-forget)
    "SIM102",  # collapsible-if (readability preference)
    "SIM105",  # suppressible-exception
    "SIM108",  # ternary-operator (readability preference)
    "W293",    # blank-line-with-whitespace
    "N802",    # function-name (property names like SQLALCHEMY_DATABASE_URI)
    "N805",    # first-arg (declared_attr uses cls)
    "N999",    # invalid-module-name (gunicorn.conf.py)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S105", "S106", "S311", "E402", "F841"]
"alembic/**/*.py" = ["E402"]
"app/core/config.py" = ["S104"]
"app/ml/**/*.py" = ["S101", "F821"]
"app/agents/**/*.py" = ["S108", "S110", "F821"]
"app/api/**/*.py" = ["S104", "E402", "UP045"]
"app/core/logging.py" = ["N806"]
"app/plugins/**/*.py" = ["B007"]
"app/services/whisper_service.py" = ["S108"]
"gunicorn.conf.py" = ["E402"]
"locustfile.py" = ["S311"]
"ml_service/**/*.py" = ["S101"]
"simulate_chaos.py" = ["S101"]